+++ src/Schema.php
@@ -17,8 +17,9 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Helper\DbArrayHelper;
+use Yiisoft\Db\Oracle\Column\BinaryColumnSchema;
 use Yiisoft\Db\Schema\Builder\ColumnInterface;
-use Yiisoft\Db\Schema\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
 use Yiisoft\Db\Schema\TableSchemaInterface;
 
 use function array_change_key_case;
@@ -405,14 +406,14 @@ final class Schema extends AbstractPdoSchema
             return false;
         }
 
-        /** @psalm-var string[][] $columns */
-        foreach ($columns as $column) {
-            /** @psalm-var ColumnInfoArray $column */
-            $column = array_change_key_case($column);
+        /** @psalm-var string[][] $info */
+        foreach ($columns as $info) {
+            /** @psalm-var ColumnInfoArray $info */
+            $info = array_change_key_case($info);
 
-            $c = $this->createColumnSchema($column);
+            $column = $this->loadColumnSchema($info);
 
-            $table->column($c->getName(), $c);
+            $table->column($info['column_name'], $column);
         }
 
         return true;
@@ -448,13 +449,21 @@ final class Schema extends AbstractPdoSchema
     }
 
     /**
-     * Creates ColumnSchema instance.
+     * Loads the column information into a {@see ColumnSchemaInterface} object.
      *
-     * @psalm-param ColumnInfoArray $info
+     * @param array $info The column information.
+     *
+     * @return ColumnSchemaInterface The column schema object.
+     *
+     * @psalm-param ColumnInfoArray $info The column information.
      */
-    protected function createColumnSchema(array $info): ColumnSchemaInterface
+    private function loadColumnSchema(array $info): ColumnSchemaInterface
     {
-        $column = new ColumnSchema($info['column_name']);
+        $dbType = $info['data_type'];
+        $type = $this->extractColumnType($dbType, $info);
+
+        $column = $this->createColumnSchema($type);
+        $column->name($info['column_name']);
         $column->allowNull($info['nullable'] === 'Y');
         $column->comment($info['column_comment']);
         $column->primaryKey((bool) $info['is_pk']);
@@ -462,12 +471,19 @@ final class Schema extends AbstractPdoSchema
         $column->size((int) $info['data_length']);
         $column->precision($info['data_precision'] !== null ? (int) $info['data_precision'] : null);
         $column->scale($info['data_scale'] !== null ? (int) $info['data_scale'] : null);
-        $column->dbType($info['data_type']);
-        $column->type($this->extractColumnType($column));
-        $column->phpType($this->getColumnPhpType($column));
+        $column->dbType($dbType);
         $column->defaultValue($this->normalizeDefaultValue($info['data_default'], $column));
 
         return $column;
+    }
+
+    protected function createColumnSchemaFromPhpType(string $phpType, string $type): ColumnSchemaInterface
+    {
+        if ($phpType === self::PHP_TYPE_RESOURCE) {
+            return new BinaryColumnSchema($type, $phpType);
+        }
+
+        return parent::createColumnSchemaFromPhpType($phpType, $type);
     }
 
     /**
@@ -645,16 +661,20 @@ final class Schema extends AbstractPdoSchema
     /**
      * Extracts the data type for the given column.
      *
-     * @param ColumnSchemaInterface $column The column schema object.
+     * @param string $dbType The database data type
+     * @param array $info Column information.
+     * @psalm-param ColumnInfoArray $info
      *
      * @return string The abstract column type.
      */
-    private function extractColumnType(ColumnSchemaInterface $column): string
+    private function extractColumnType(string $dbType, array $info): string
     {
-        $dbType = strtolower((string) $column->getDbType());
+        $dbType = strtolower($dbType);
 
         if ($dbType === 'number') {
-            return match ($column->getScale()) {
+            $scale = $info['data_scale'] !== null ? (int) $info['data_scale'] : null;
+
+            return match ($scale) {
                 null => self::TYPE_DOUBLE,
                 0 => self::TYPE_INTEGER,
                 default => self::TYPE_DECIMAL,
@@ -663,7 +683,7 @@ final class Schema extends AbstractPdoSchema
 
         $dbType = preg_replace('/\([^)]+\)/', '', $dbType);
 
-        if ($dbType === 'interval day to second' && $column->getPrecision() > 0) {
+        if ($dbType === 'interval day to second' && $info['data_precision'] > 0) {
             return self::TYPE_STRING;
         }