+++ tests/ColumnSchemaTest.php
@@ -4,17 +4,23 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Oracle\Tests;
 
-use PHPUnit\Framework\TestCase;
+use PDO;
+use Yiisoft\Db\Command\Param;
 use Yiisoft\Db\Expression\Expression;
+use Yiisoft\Db\Oracle\Column\BinaryColumnSchema;
 use Yiisoft\Db\Oracle\Tests\Support\TestTrait;
 use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\Schema\Column\DoubleColumnSchema;
+use Yiisoft\Db\Schema\Column\IntegerColumnSchema;
+use Yiisoft\Db\Schema\Column\StringColumnSchema;
+use Yiisoft\Db\Tests\Common\CommonColumnSchemaTest;
 
 use function str_repeat;
 
 /**
  * @group oracle
  */
-final class ColumnSchemaTest extends TestCase
+final class ColumnSchemaTest extends CommonColumnSchemaTest
 {
     use TestTrait;
 
@@ -61,5 +67,41 @@ final class ColumnSchemaTest extends TestCase
         $this->assertSame(0b0110_0110, $bitColPhpType);
 
         $db->close();
+    }
+
+    public function testColumnSchemaInstance()
+    {
+        $db = $this->getConnection(true);
+        $schema = $db->getSchema();
+        $tableSchema = $schema->getTableSchema('type');
+
+        $this->assertInstanceOf(IntegerColumnSchema::class, $tableSchema->getColumn('int_col'));
+        $this->assertInstanceOf(StringColumnSchema::class, $tableSchema->getColumn('char_col'));
+        $this->assertInstanceOf(DoubleColumnSchema::class, $tableSchema->getColumn('float_col'));
+        $this->assertInstanceOf(BinaryColumnSchema::class, $tableSchema->getColumn('blob_col'));
+    }
+
+    /** @dataProvider \Yiisoft\Db\Oracle\Tests\Provider\ColumnSchemaProvider::predefinedTypes */
+    public function testPredefinedType(string $className, string $type, string $phpType)
+    {
+        parent::testPredefinedType($className, $type, $phpType);
+    }
+
+    /** @dataProvider \Yiisoft\Db\Oracle\Tests\Provider\ColumnSchemaProvider::dbTypecastColumns */
+    public function testDbTypecastColumns(string $className, array $values)
+    {
+        parent::testDbTypecastColumns($className, $values);
+    }
+
+    public function testBinaryColumnSchema()
+    {
+        $binaryCol = new BinaryColumnSchema();
+        $binaryCol->dbType('BLOB');
+
+        $this->assertInstanceOf(Expression::class, $binaryCol->dbTypecast("\x10\x11\x12"));
+        $this->assertInstanceOf(
+            Expression::class,
+            $binaryCol->dbTypecast(new Param("\x10\x11\x12", PDO::PARAM_LOB)),
+        );
     }
 }